Move the trace mark/measurement classes to the trace plugin package as they are mostly used there; Rename TracePerformanceEntry to PerformanceEntry.
