+++ src/test/java/com/iopipe/__DoPluginTest__.java
@@ -2,9 +2,6 @@ package com.iopipe;
 
 import com.iopipe.http.RemoteRequest;
 import com.iopipe.http.RemoteResult;
-import com.iopipe.plugin.trace.TraceExecution;
-import com.iopipe.plugin.trace.TracePlugin;
-import com.iopipe.TraceMeasurement;
 import java.util.Map;
 import javax.json.JsonObject;
 import javax.json.JsonString;
