+++ src/test/java/com/iopipe/__DoTracePlugin__.java
@@ -3,8 +3,8 @@ package com.iopipe;
 import com.iopipe.http.RemoteRequest;
 import com.iopipe.http.RemoteResult;
 import com.iopipe.plugin.trace.TraceExecution;
+import com.iopipe.plugin.trace.TraceMeasurement;
 import com.iopipe.plugin.trace.TracePlugin;
-import com.iopipe.TraceMeasurement;
 import java.util.Map;
 import javax.json.JsonObject;
 import javax.json.JsonString;
